Share combine-function KDoc example rendering
The KDoc example generation for combine functions was duplicated three
ways: the `[A] + [B] -> [Target]` auto-description line and the
`val a = A(...) / val target = a.fn(...)` example body appeared in
CombineToClass.kt's Basic and Override examples and again in Object.kt's
Basic example.

Extract `appendCombineAutoDescription` (a KDocWriter extension, so the
` * ` line prefix is preserved) and `combineExampleBody(sources, funName,
extraArgs)` into transform/CombineExample.kt, and call them from both
class- and object-target KDoc builders. The override example reuses the
same body with `extraArgs = ", property = value"`. CombineToClass.kt drops
from 302 to 249 lines.

Behavior unchanged (snapshot suite byte-identical).
